restructure(refactor)- reated page specific folder
