+++ src/components/pages/instructions/index.jsx
@@ -1,6 +1,7 @@
-import React from 'react'
+import { useSelector } from 'react-redux'
 
 const InstructionsCatgory = () => {
+
   return (
     <div>InstructionsCatgory</div>
   )
